requirements.txt
Django==4.0.6
mysql-connector-python==8.0.29
cassandra-driver==3.25.0
gunicorn
